parallelization of new code search
Parallelized the process of searching new codes. New codes get searched by iterating over l, m that create the check matrices.
The polynomials for A and B are set fixed to 3 terms with one x and two y terms for A and one y and two x terms.
All combinations will be tested and controlled for code validity and higher
new encoding rate that 1/15. All results will be saved to a list of dictionaries with
the dicts saving the code parameters.